fix integer size issue in ContentTypes.cpp which can lead to an infinite loop

+++ libresapi/src/util/ContentTypes.cpp
@@ -43,8 +43,8 @@ std::string ContentTypes::cTypeFromExt(const std::string &extension)
 	while(getline(file, line))
 	{
 		if(line.empty() || line[0] == '#') continue;
-		unsigned int i = line.find_first_of("\t ");
-		unsigned int j;
+        size_t i = line.find_first_of("\t ");
+        size_t j;
 		while(i != std::string::npos)	//tokenize
 		{
 			j = i;
